Created a API Connector to follow the single responsibility principle. Implemented the functions as trait to mock them later
