+++ Cargo.toml
@@ -33,6 +33,7 @@ webbrowser = "^0"
 fs_extra = "^1"
 globset = "^0"
 chrono = "^0"
+async-trait = "^0"
 
 [build-dependencies]
 embed-resource = "^1"
