+++ src/core/mod.rs
@@ -1,11 +1,13 @@
+pub mod api_connector;
 pub mod config;
 pub mod installer;
 pub mod plugin;
 pub mod plugin_parser;
 pub mod synchronizer;
 
+pub use api_connector::APIConnector;
 pub use config::Config;
 pub use installer::Installer;
-pub use plugin::{Base, Plugin, Installed};
+pub use plugin::{Base, Installed, Plugin};
 pub use plugin_parser::PluginParser;
 pub use synchronizer::Synchronizer;
